--- Assets/Remnants/Materials/PreFabs/Bee Enemey.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3b2794a224993cc4ba63b6411c3d31f3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
