--- Assets/Remnants/Materials/PreFabs/Bullet.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 50134238e27764c40ba45229638049aa
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
