--- Assets/Remnants/Materials/PreFabs/Ruby.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9dd0a744590d78d4187af80b93d69b4f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
