+++ Assets/Remnants/PreFabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 977673355f0aee14f9be00e123f28903
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
