--- Assets/Remnants/Remnant.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b6e861dd8fffbe44abf921d784f0327b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
