--- Assets/Remnants/Remnant/GameSettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 53b6e12c5d721d34786641a4497db0f4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
