--- Assets/Remnants/Remnant/HomingBullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca3d3168b0aa3bd49959209e2f3f7d27
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
