--- Assets/Remnants/Remnant/Posion.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5d85a6b184d07b347841ab17c77301f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
