--- Assets/Remnants/Remnant/Stinger animation.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: da9fff7ad669841478926d00ea4c7be0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
