+++ Assets/Remnants/Ruby Gameplay/DragonFollow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 37ca7a7a05ee8c54a8b587e1651b2c8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
