+++ Assets/Remnants/Ruby Gameplay/EnemyFollowPlayer.cs
@@ -54,4 +54,11 @@ public class EnemyFollowPlayer : MonoBehaviour
         Gizmos.DrawWireSphere(transform.position, lineOfSite);
         Gizmos.DrawWireSphere(transform.position, shootingRange);
     }
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+        if (collision.CompareTag("Fire"))
+        {
+            Destroy(gameObject);
+        }
+    }
 }
