+++ Assets/Remnants/Ruby Gameplay/Remnant.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6e861dd8fffbe44abf921d784f0327b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
