+++ Assets/Remnants/Ruby Gameplay/Remnant/GameSettings.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class GameSettings : MonoBehaviour
+{
+    //Holds the name of the next level 
+    public string nextLevelName;
+
+    //Used by the button to load the scene 
+    public void LoadScene()
+    {
+        SceneManager.LoadScene(nextLevelName);
+    }
+}
+
+    //Used by a Trigger Collider 
+
