+++ Assets/Remnants/Ruby Gameplay/Remnant/Shoot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0bf4214d35b862a4aa21f25095405089
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
