+++ Assets/Remnants/Ruby Gameplay/Remnant/SpiderFollow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 805d4b65c5d49a645a9fcb06b6f806a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
