+++ Assets/Remnants/Ruby Gameplay/Remnant/Stinger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15817b2157976094595ad07553b4e64d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
