+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - TutorialRequirement
   - Bee
   - Spider
+  - Fire
   layers:
   - Default
   - TransparentFX
