+++ UserSettings/EditorUserSettings.asset
@@ -12,16 +12,16 @@ EditorUserSettings:
       value: 050751035105505a590d0f7346775c4413161e7279297266287e1e66e4b4616f
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 56030c5654030b0c085f0f7615275c4412154a2c2e7b70362f284b67b7b3376c
+      value: 520006025d015c5f585b5f2645220d444f4f4e7874717f647c2c1e6ab1e2633a
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 520006025d015c5f585b5f2645220d444f4f4e7874717f647c2c1e6ab1e2633a
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     RecentlyUsedSceneGuid-4:
       value: 5606565603005c0a5c0d597114750a4445154d7d2f7c7f32287b4935b7b4643d
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      value: 56030c5654030b0c085f0f7615275c4412154a2c2e7b70362f284b67b7b3376c
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
